sql cte subquery partition
An extension to the original query where we aggregated our results with a group by clause to include min, max, and average. By aggregating our data we lose some granularity. However, if one wanted to keep the granularity and in this case all the patients encounters and Blood Urea Nitrogen (BUN) values and also keep the min, max, and average columns in the table then a partition function would be used.